remove redshift_helper and add rs_command to glue job script
